+++ src/index.js
@@ -78,3 +78,6 @@ initSidebar = require('./viewer/initSidebar');
 HoverMenuItem = require('./stuff/HoverMenuItem');
 HoverMenu = require('./stuff/HoverMenu');
 Potree.GLProgram = require('./webgl/GLProgram');
+Potree.updatePointClouds = require('./util/updatePointClouds');
+Potree.updateVisibility = require('./util/updateVisibility');
+Potree.updateVisibilityStructures = require('./util/updateVisibilityStructures');
